$ refactor go code for package convention
