+++ go/controller/ping_controller/ping_controller.go
@@ -2,8 +2,7 @@ package ping_controller
 
 import (
 	"github.com/graphql-go/graphql"
-
-	ping_service "github.com/jswildcards/gocareer/service"
+	"github.com/jswildcards/gocareer/service/ping_service"
 )
 
 func Ping(p graphql.ResolveParams) (interface{}, error) {
